requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.3.0
colorama==0.4.6
coverage==7.10.7
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.118.3
flake8==7.3.0
greenlet==3.2.4
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
mccabe==0.7.0
packaging==25.0
pluggy==1.6.0
psycopg[binary]>=3.1 
pycodestyle==2.14.0
pydantic==2.12.0
pydantic_core==2.41.1
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.2
pytest-cov==7.0.0
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
